fix: ledger-correct thirdparty.go, capture_linux.go, and migration 00090
Fixed ledger violations discovered during audit of PR #49 code:

1. thirdparty.go GetPackages: rows.Scan error used continue instead of
   returning error. rows.Err() not checked after iteration.
   Ledger rule: Never ignore scan errors; check rows.Err() after iteration.

2. capture_linux.go gradient fallback: integer overflow in uint8 conversion
   (x * 255 / width). gosec G115 HIGH severity.
   Fix: use float64 intermediate calculation.

3. migration 00090: Missing FORCE ROW LEVEL SECURITY.
   Ledger rule: RLS tables require ENABLE + FORCE.

4. pkg/postgres/schema.go: Missing embedded migration 90 for parity.
   Ledger rule: Embedded and standalone migrations must be mechanically reconciled.

+++ internal/remote/capture_linux.go
@@ -157,8 +157,8 @@ func createGradientFallback(width, height int) image.Image {
 	img := image.NewRGBA(image.Rect(0, 0, width, height))
 	for y := 0; y < height; y++ {
 		for x := 0; x < width; x++ {
-			r := uint8((x * 255) / width)
-			g := uint8((y * 255) / height)
+			r := uint8(float64(x) * 255.0 / float64(width))
+			g := uint8(float64(y) * 255.0 / float64(height))
 			b := uint8(128)
 			img.Set(x, y, color.RGBA{R: r, G: g, B: b, A: 255})
 		}
